USER REVIEW ON OWN DEVELOPED WEBSITE:AN EXPERIMENTAL STUDY WITH CERTIFICATE OF PAPER
This paper aims to provide reviews to the buyers who wants to buy the product online. The
feedback reviews form the previous users, who have used the same product will make the
buying process easy. This website will help them to take feedback for the product either
online or offline at same place. The problem is with the current scenario in the feedback
system are like as lack of security ,accessing the data, authenticity of the feedback, no review
availability etc. Nowadays, all the products are not available at the same place they are
available on different websites, which is inconvenient and time-consuming process to check
the maximum possible feedback about the product from different websites.